# openapi/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Spring Application exposes Api to manage configuration for EC/PSP on the Nodo dei Pagamenti",
    "termsOfService": "https://www.pagopa.gov.it/",
    "title": "PagoPA API configuration",
    "version": "0.2.28"
  },
  "host": "127.0.0.1:8080",
  "basePath": "/apiconfig/api/v1",
  "schemes": [
    "http"
  ],
  "paths": {
    "/brokers": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "default": 50,
            "description": "Number of elements on one page. Default = 50",
            "format": "int32",
            "in": "query",
            "name": "limit",
            "required": false,
            "type": "integer"
          },
          {
            "description": "Page number. Page value starts from 0",
            "format": "int32",
            "in": "query",
            "name": "page",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Brokers"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Creditor Institutions"
        ],
        "operationId": "getBrokers",
        "summary": "Get paginated list of creditor brokers"
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BrokerDetails"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/BrokerDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Creditor Institutions"
        ],
        "operationId": "createBroker",
        "summary": "Create a broker"
      }
    },
    "/brokers/{brokercode}": {
      "delete": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "broker code",
            "in": "path",
            "maxLength": 50,
            "minLength": 0,
            "name": "brokercode",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Creditor Institutions"
        ],
        "operationId": "deleteBroker",
        "summary": "Delete a broker"
      },
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "broker code.",
            "in": "path",
            "maxLength": 50,
            "minLength": 0,
            "name": "brokercode",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BrokerDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Creditor Institutions"
        ],
        "operationId": "getBroker",
        "summary": "Get creditor broker details "
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "broker code",
            "in": "path",
            "maxLength": 50,
            "minLength": 0,
            "name": "brokercode",
            "required": true,
            "type": "string"
          },
          {
            "description": "The values to update of the broker",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BrokerDetails"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BrokerDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Creditor Institutions"
        ],
        "operationId": "updateBroker",
        "summary": "Update a broker"
      }
    },
    "/brokerspsp": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "default": 50,
            "description": "Number of elements on one page. Default = 50",
            "format": "int32",
            "in": "query",
            "name": "limit",
            "required": false,
            "type": "integer"
          },
          {
            "description": "Page number. Page value starts from 0",
            "format": "int32",
            "in": "query",
            "name": "page",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BrokersPsp"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Payment Service Providers"
        ],
        "operationId": "getBrokersPsp",
        "summary": "Get paginated list of PSP brokers"
      }
    },
    "/brokerspsp/{brokerpspcode}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Broker code of a PSP.",
            "in": "path",
            "maxLength": 50,
            "minLength": 0,
            "name": "brokerpspcode",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BrokerPspDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Payment Service Providers"
        ],
        "operationId": "getBrokerPsp",
        "summary": "Get creditor broker details "
      }
    },
...
  },
  "definitions": {
    "Broker": {
      "properties": {
        "broker_code": {
          "example": "223344556677889900",
          "maxLength": 35,
          "minLength": 0,
          "type": "string"
        },
        "description": {
          "example": "Lorem ipsum dolor sit amet",
          "maxLength": 255,
          "minLength": 0,
          "type": "string"
        },
        "enabled": {
          "type": "boolean"
        }
      },
      "required": [
        "broker_code",
        "description",
        "enabled"
      ],
      "type": "object"
    },
    "BrokerDetails": {
      "properties": {
        "broker_code": {
          "example": "223344556677889900",
          "maxLength": 35,
          "minLength": 0,
          "type": "string"
        },
        "description": {
          "example": "Lorem ipsum dolor sit amet",
          "maxLength": 255,
          "minLength": 0,
          "type": "string"
        },
        "enabled": {
          "type": "boolean"
        },
        "extended_fault_bean": {
          "type": "boolean"
        }
      },
      "required": [
        "broker_code",
        "description",
        "enabled",
        "extended_fault_bean"
      ],
      "type": "object"
    },
    "BrokerPsp": {
      "properties": {
        "broker_psp_code": {
          "example": "223344556677889900",
          "maxLength": 35,
          "minLength": 0,
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "enabled": {
          "type": "boolean"
        }
      },
      "required": [
        "broker_psp_code",
        "description",
        "enabled"
      ],
      "type": "object"
    },
    "BrokerPspDetails": {
      "properties": {
        "broker_psp_code": {
          "example": "223344556677889900",
          "maxLength": 35,
          "minLength": 0,
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "enabled": {
          "type": "boolean"
        },
        "extended_fault_bean": {
          "type": "boolean"
        }
      },
      "required": [
        "broker_psp_code",
        "description",
        "enabled",
        "extended_fault_bean"
      ],
      "type": "object"
    },
    "Brokers": {
      "properties": {
        "brokers_list": {
          "items": {
            "$ref": "#/definitions/Broker"
          },
          "type": "array"
        },
        "page_info": {
          "$ref": "#/definitions/PageInfo"
        }
      },
      "required": [
        "brokers_list",
        "page_info"
      ],
      "type": "object"
    },
    "BrokersPsp": {
      "properties": {
        "brokers_psp_list": {
          "items": {
            "$ref": "#/definitions/BrokerPsp"
          },
          "type": "array"
        },
        "page_info": {
          "$ref": "#/definitions/PageInfo"
        }
      },
      "required": [
        "brokers_psp_list",
        "page_info"
      ],
      "type": "object"
    },
...
    "PageInfo": {
      "properties": {
        "items_found": {
          "description": "Number of items found. (The last page may have fewer elements than required)",
          "format": "int32",
          "type": "integer"
        },
        "limit": {
          "description": "Required number of items per page",
          "format": "int32",
          "type": "integer"
        },
        "page": {
          "description": "Page number",
          "format": "int32",
          "type": "integer"
        },
        "total_pages": {
          "description": "Total number of pages",
          "format": "int32",
          "type": "integer"
        }
      },
      "required": [
        "items_found",
        "limit",
        "page",
        "total_pages"
      ],
      "type": "object"
    },
...
    "ProblemJson": {
      "properties": {
        "detail": {
          "description": "A human readable explanation specific to this occurrence of the problem.",
          "example": "There was an error processing the request",
          "type": "string"
        },
        "status": {
          "description": "The HTTP status code generated by the origin server for this occurrence of the problem.",
          "example": 200,
          "format": "int32",
          "maximum": 600,
          "minimum": 100,
          "type": "integer"
        },
        "title": {
          "description": "A short, summary of the problem type. Written in english and readable for engineers (usually not suited for non technical stakeholders and not localized); example: Service Unavailable",
          "type": "string"
        }
      },
      "type": "object"
    },
...
  },
  "securityDefinitions": {
    "ApiKey": {
      "description": "The API key to access this function app.",
      "in": "header",
      "name": "Ocp-Apim-Subscription-Key",
      "type": "apiKey"
    },
    "Authorization": {
      "bearerFormat": "JWT",
      "description": "JWT token get after Azure Login",
      "scheme": "Bearer",
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
...
}